+++ client/src/Component/MiniNav/MiniNav.jsx
@@ -1,6 +1,7 @@
 import React from "react"
-import styles from './MiniNav.module.css'
 import { Link } from "react-router-dom"
+
+import styles from './MiniNav.module.css'
 
 const MiniNav = () => {
