+++ client/src/Component/ModalRegister/ModalRegister.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import Button from 'react-bootstrap/Button';
 import Modal from 'react-bootstrap/Modal';
+
 import styles from './ModalRegister.module.css'
 
 function ModalRegister(props) {
