+++ client/src/Component/NavBar/NavBar.jsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { useDispatch } from 'react-redux';
 import { Link, useHistory, useLocation } from "react-router-dom";
 
-import { getProductsByQuery } from '../../Redux/Action/index.js'
+import { getProductsByQuery } from '../../Redux/Actions/products.js'
 
 import Button from '../Button/Button.jsx';
 import Input from "../Input/Input.jsx";
