+++ client/src/Component/Slider/Slider.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import Carousel from 'react-bootstrap/Carousel'
+
 import styles from './Slider.module.css'
 
 export default function Slider() {
