+++ client/src/hooks/ObjectToQuery.js
@@ -1,5 +1,3 @@
-
-
 export const objectToQuery = (query) =>{
     let arrayObjeto = Object.entries(query) // pongo el objeto en formato array tipo [[Propiedad1,valor propiedad1],[Propiedad2,Valor2]] etc            
     let stringToUrl = ''
